+++ README.md
@@ -1,1 +1,4 @@
 Item Details Module
+npm run react-dev
+npm run server-dev
+npm run db:setup
